chore(deps): use cross-fetch instead of node-fetch
