+++ frontend/package.json
@@ -6,7 +6,6 @@
   "bugs": "https://github.com/npocccties/chiloportal/issues",
   "dependencies": {
     "@aspida/fetch": "^1.14.0",
-    "@aspida/node-fetch": "^1.14.0",
     "@faker-js/faker": "^8.4.1",
     "@headlessui/react": "^1.7.18",
     "@iconify/react": "^4.1.1",
@@ -14,13 +13,13 @@
     "ajv-formats": "^2.1.1",
     "aspida": "^1.14.0",
     "clsx": "^2.1.0",
+    "cross-fetch": "^4.0.0",
     "fast-glob": "^3.3.2",
     "image-size": "^1.1.1",
     "jotai": "^2.7.0",
     "msw": "^2.2.2",
     "next": "^14.1.1",
     "next-mdx-remote": "^4.4.1",
-    "node-fetch": "^2.7.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-textarea-autosize": "^8.5.3",
@@ -39,7 +38,6 @@
     "@svgr/webpack": "^8.1.0",
     "@tailwindcss/typography": "^0.5.10",
     "@types/node": "^20.11.24",
-    "@types/node-fetch": "^2.6.11",
     "@types/react": "^18.2.61",
     "@types/react-dom": "^18.2.19",
     "@typescript-eslint/eslint-plugin": "^7.1.0",
