+++ frontend/pages/search.tsx
@@ -1,7 +1,7 @@
 import { GetServerSidePropsResult } from "next";
 import Error from "next/error";
 import Head from "next/head";
-import { client, getErrorProps } from "lib/node-client";
+import { client, getErrorProps } from "lib/client";
 import Template from "templates/Search";
 import { pagesPath } from "lib/$path";
 import title from "lib/title";
